feat: improved async tests
