+++ spec/helper/asyncUtilities.js
@@ -1,0 +1,20 @@
+export async function sleep(duration) {
+    return new Promise((resolve) => setTimeout(() => resolve(), duration));
+}
+
+export async function waitFor(timeoutDuration, checkFunction) {
+    return new Promise((resolve, reject) => {
+        let checkInterval;
+        const waitTimeout = setTimeout(() => {
+            clearInterval(checkInterval);
+            reject();
+        }, timeoutDuration);
+        checkInterval = setInterval(() => {
+            if (checkFunction()) {
+                clearTimeout(waitTimeout);
+                clearInterval(checkInterval);
+                resolve();
+            }
+        }, 100);
+    });
+}
